Cleanup: move relative time messaging code to its own module
